@@ -26,22 +26,22 @@
     "watch": "nodemon --exec 'npx gulp buildTmLanguage'"
   },
   "devDependencies": {
-    "@types/chai": "4.2.19",
-    "@types/gulp": "4.0.8",
-    "@types/mocha": "8.2.2",
-    "chai": "4.3.4",
-    "gulp": "4.0.2",
-    "gulp-json2cson": "2.0.0",
-    "gulp-mocha": "8.0.0",
-    "gulp-typescript": "6.0.0-alpha.1",
-    "gulp-yaml": "2.0.4",
-    "js-yaml": "4.1.0",
-    "mocha": "8.3.2",
-    "nodemon": "2.0.20",
-    "plist": "3.0.5",
-    "ts-node": "10.0.0",
-    "typescript": "4.3.5",
-    "vscode-oniguruma": "1.5.1",
-    "vscode-textmate": "5.4.0"
+    "@types/chai": "^4.3.20",
+    "@types/gulp": "^4.0.17",
+    "@types/mocha": "^10.0.10",
+    "chai": "^6.0.1",
+    "gulp": "^5.0.1",
+    "gulp-json2cson": "^2.0.0",
+    "gulp-mocha": "^10.0.1",
+    "gulp-typescript": "^6.0.0-alpha.1",
+    "gulp-yaml": "^2.0.4",
+    "js-yaml": "^4.1.0",
+    "mocha": "^11.7.1",
+    "nodemon": "^3.1.10",
+    "plist": "^3.1.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.9.2",
+    "vscode-oniguruma": "^2.0.1",
+    "vscode-textmate": "^9.2.0"
   }
 }
